+++ libinstpatch/IpatchVBank.c
@@ -339,8 +339,8 @@ ipatch_vbank_base_find_unused_locale (IpatchBase *base, int *bank,
     lbank = lprogram >> 16;
     lprogram &= 0xFFFF;
 
-    if (lbank > b || (lbank == b && lprogram > n)) break;
-    if (lbank >= b)
+    if (lbank > (guint)b || (lbank == b && lprogram > (guint)n)) break;
+    if (lbank >= (guint)b)
     {
       if (++n > 127)
       {
